remove non rn compatible extensions
