--- packages/create/src/components/FlashList/index.web.tv.tsx
@@ -1,3 +1,0 @@
-import FlashList from './index.native.tv';
-
-export default FlashList;
