--- packages/create/src/components/ImageBackground/index.web.tv.tsx
@@ -1,3 +1,0 @@
-import ImageBackground from './index.native.tv';
-
-export default ImageBackground;
