--- packages/create/src/components/Screen/index.web.tv.tsx
@@ -1,3 +1,0 @@
-import Screen from './index.native.tv';
-
-export default Screen;
