--- packages/create/src/components/ScrollView/index.web.tv.tsx
@@ -1,3 +1,0 @@
-import ScrollView from './index.native.tv';
-
-export default ScrollView;
