--- packages/create/src/components/TouchableOpacity/index.native.tv.tsx
@@ -1,3 +1,0 @@
-import TouchableOpacity from '../Pressable';
-
-export default TouchableOpacity;
