--- packages/create/src/components/TouchableOpacity/index.web.tv.tsx
@@ -1,3 +1,0 @@
-import TouchableOpacity from '../Pressable';
-
-export default TouchableOpacity;
