--- packages/create/src/components/View/index.native.tv.tsx
@@ -1,25 +1,0 @@
-import React from 'react';
-import { View as RNView } from 'react-native';
-import Pressable from '../Pressable';
-import type { ViewProps } from '../../focusManager/types';
-import ViewGroup from '../ViewGroup';
-
-const View = React.forwardRef<RNView, ViewProps>(({ children, focusContext, focusOptions, ...props }, ref) => {
-    if (focusOptions?.group) {
-        return (
-            <ViewGroup focusContext={focusContext} focusOptions={focusOptions} {...props} ref={ref}>
-                {children}
-            </ViewGroup>
-        );
-    }
-
-    return (
-        <Pressable focusContext={focusContext} {...props} focus={false} ref={ref}>
-            {children}
-        </Pressable>
-    );
-});
-
-View.displayName = 'View';
-
-export default View;
