+++ packages/create/src/focusManager/service/keyHandler.web.ts
@@ -45,6 +45,7 @@ class KeyHandler {
 
     private onKeyDown(eventType: string) {
         if (CoreManager.isFocusManagerEnabled()) {
+            console.log({ eventType });
             if (eventType === 'playPause') {
                 CoreManager.setDebuggerEnabled(!CoreManager.isDebuggerEnabled());
                 Logger.setIsDebuggerEnabled(CoreManager.isDebuggerEnabled()).debug(CoreManager);
