+++ packages/create/src/focusableView/index.tsx
@@ -1,31 +1,13 @@
-import React from 'react';
-import { View as RNView, ViewProps, ColorValue } from 'react-native';
+import { requireNativeComponent, ViewProps, ViewStyle, HostComponent } from 'react-native';
 import { AnimatorBackground, AnimatorBorder, AnimatorScale, AnimatorScaleWithBorder } from '..';
 
-// AnimatorBackground | AnimatorBorder | AnimatorScale | AnimatorScaleWithBorder
-interface Props extends ViewProps {
+interface TVFocusableView extends ViewProps {
     animatorOptions?: {
-        type:
-            | AnimatorBackground['type']
-            | AnimatorBorder['type']
-            | AnimatorScale['type']
-            | AnimatorScaleWithBorder['type'];
-        focus:
-            | AnimatorBackground['focus']
-            | AnimatorBorder['focus']
-            | AnimatorScale['focus']
-            | AnimatorScaleWithBorder['focus'];
-        blur: {
-            borderWidth?: number;
-            duration?: number;
-            borderRadius?: number;
-            borderColor?: ColorValue;
-            backgroundColor?: ColorValue;
-        };
-        // style: ViewStyle;
+        animator: AnimatorBackground | AnimatorBorder | AnimatorScale | AnimatorScaleWithBorder;
+        style: ViewStyle;
     };
 }
 
-const View = React.forwardRef<RNView, Props>((props, ref) => <RNView {...props} ref={ref} />);
+const TvFocusableViewManager: HostComponent<TVFocusableView> = requireNativeComponent('TvFocusableView');
 
-export default View;
+export default TvFocusableViewManager;
