--- packages/create/src/remoteHandler/index.firetv.tsx
@@ -1,1 +1,0 @@
-export { useTVRemoteHandler } from './index.androidtv';
