+++ packages/template-starter/src/pages/_app.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { View } from '@flexn/create';
+import { Debugger, App } from '@flexn/create';
 import Menu from '../components/menu';
 import { themeStyles, ThemeProvider } from '../config';
 import { AppProps } from 'next/app';
@@ -8,9 +8,10 @@ export default function MyApp({ Component, pageProps }: AppProps<any>) {
     return (
         <ThemeProvider>
             <Menu />
-            <View style={themeStyles.appContainer}>
+            <App style={themeStyles.appContainer}>
                 <Component {...pageProps} />
-            </View>
+                <Debugger />
+            </App>
         </ThemeProvider>
     );
 }
